feat: separate commit message config (#68)
* feat: separate commit message config

The PR title in many of my own cases includes a "draft" tag
that's meant to be removed.
This ends up as the commit message in the end,
which is not so easily changed.
This separates the PR title and commit message
into different options.

* docs: add new config option

* chore: version bump to 2.1.0

* chore: CI fix

* chore: CI fixes

* chore: the tiniest CI fix

This is what I get editing on mobile

* fix: remove weird end char

I had copied some output from GH Actions logs
to produce the precommit fixes seen here.

In hindsight, I should generalize the autofix method
I pushed in a separate project,
so the pre-commit checks can commit their changes.